// frontend/src/generated/CrmMetadataResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CrmPipeline } from "./CrmPipeline";
import type { CrmPropertyDefinition } from "./CrmPropertyDefinition";

export type CrmMetadataResponse = { properties: Array<CrmPropertyDefinition>, pipelines: Array<CrmPipeline>, fetched_at: string, hubspot_ms: number, total_ms: number, cache_hit: boolean, };